[refactor]:  SearchCateogryDotButton 클래스 ResultSearchViewController 내부 클래스로 구현 통해 해당 파일 사용 범위 제한
