+++ Zatch/domain/ViewControllers/Search/ResultSearchViewController.swift
@@ -17,6 +17,22 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
     
     //MARK: UIComponent
     
+    class SearchCateogryDotButton: UIButton {
+        
+        //MARK: - LifeCycle
+        override init(frame: CGRect) {
+            super.init(frame: .zero)
+            
+            self.setImage(UIImage(named: "search_dot"), for: .normal)
+            self.setImage(UIImage(named: "search_dot_checked"), for: .selected)
+        }
+        
+        required init?(coder: NSCoder) {
+            fatalError("init(coder:) has not been implemented")
+        }
+        
+    }
+    
     let topView = UIView()
     
     let separateLine = UIView().then{
